remove unused vars and make constants file
