+++ src/server/items/models.js
@@ -1,11 +1,7 @@
 import db from '../database';
-import {KeyUtils, STATE, CATEGORY, DEFAULT_PRECISON} from '../key-utils';
+import {KeyUtils} from '../key-utils';
+import {STATE, CATEGORY, DEFAULT_PRECISON, STATE_STRING} from '../constants';
 
-export const STATE_STRING = {
-  [STATE.ALIVE]: 'alive',
-  [STATE.EXPIRED]: 'expired',
-  [STATE.REMOVED]: 'removed'
-};
 export default class ItemManager {
   static validChecker(item, cb) {
     if (item.category === CATEGORY.FACILITY || this.isValid(item.endTime)) {
